+++ src/main/java/ru/vyarus/spock/jupiter/engine/ExtensionUtils.java
@@ -89,6 +89,11 @@ public final class ExtensionUtils {
     public static final List<Class<? extends Extension>> UNSUPPORTED_EXTENSIONS = Arrays.asList(
             // not included because it doesn't matter in context of spock
             TestTemplateInvocationContextProvider.class,
+            // not enabled to avoid class not found exception when used with 5.12
+            // (templates are obviously not supported)
+            // ClassTemplateInvocationContextProvider.class,
+            // BeforeClassTemplateInvocationCallback.class,
+            // AfterClassTemplateInvocationCallback.class,
 
             // impossible to add (spock does not allow this)
             TestInstanceFactory.class,